Closes #7 : Mostly every functions have been moved in geometry.py
